Archiving the blueprint functionality.
