+++ Archive/README.md
@@ -1,0 +1,2 @@
+### What is this directory?
+We are storing the old blueprint library code here until we can do a more thorough review and refactor of the original functionality.
