--- OSVRUnreal/Plugins/OSVR/Source/OSVR/Private/OSVRInterfaceCollection.h
@@ -1,66 +1,0 @@
-//
-// Copyright 2014, 2015 Razer Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-#pragma once
-
-#include "OSVRInterface.h"
-#include <functional>
-
-class OSVRInterfaceCollection
-{
-public:
-	OSVRInterfaceCollection(OSVR_ClientContext osvrClientContext);
-
-	OSVRInterface* GetInterface(const FName& Name)
-	{
-		return GetInterfaceImpl(Name, false);
-	}
-	OSVRInterface* GetOrCreateInterface(const FName& Name)
-	{
-		return GetInterfaceImpl(Name, true);
-	}
-
-	struct RegistrationToken
-	{
-		static const int32 INVALID_TOKEN = -1;
-
-		RegistrationToken()
-			: Token(INVALID_TOKEN)
-		{
-		}
-		explicit RegistrationToken(int32 InToken)
-			: Token(InToken)
-		{
-		}
-
-		int32 Token;
-	};
-
-	RegistrationToken RegisterOnStateChangedCallback(OSVRInterface::StateChangedCallback Function);
-	void DeregisterOnStateChangedCallback(RegistrationToken Token);
-
-private:
-	void AddDefaultInterfaces();
-	void ClearAllInterfaces();
-
-	OSVRInterface* GetInterfaceImpl(const FName& Name, bool CreateIfNotExist);
-
-	void PropagateCallbacks(OSVRInterface* Interface, uint32 Caps);
-
-	TArray< OSVRInterface::StateChangedCallback > Callbacks;
-	TArray< TSharedPtr< OSVRInterface > > Interfaces;
-	OSVR_ClientContext osvrClientContext;
-};
